requirements.txt
pytest
requests
python-dotenv
pytest-html
allure-pytest